two context managers as fixtures
